Risk Assessment/requirements.txt
Flask==3.0.3
openpyxl==3.1.5
waitress==3.0.0
Authlib==1.6.6
python-dotenv==1.2.1
requests==2.32.3
